requirements.txt
langchain
ipykernel
beautifulsoup4
python-dotenv
langchain_community
groq
langchain_text_splitters
langchain_huggingface
faiss-cpu
langserve 
fastapi
uvicorn
langchain_core